# src/ComicNew.Api/swagger.yaml
openapi: 3.0.1
info:
  title: ComicNew API
  version: 1.0.0
  description: |
    Tài liệu OpenAPI mô tả các endpoint chính của ComicNew API để tích hợp frontend dễ dàng.
    
    Luồng đăng nhập (tóm tắt):
    1. Frontend nhận token (JWT) từ Supabase.
    2. Frontend gửi token trong header `Authorization: Bearer <token>` tới API.
    3. API dùng JwtBearer middleware để xác thực token và gọi `OnTokenValidated` để đồng bộ user vào DB.
    4. Sau xác thực, các endpoint bảo vệ bằng `Authorize` có thể truy cập `HttpContext.User` để lấy claims.
servers:
  - url: http://localhost:5000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReadingHistoryResponse:
      type: object
      properties:
        Id:
          type: string
          format: uuid
        UserId:
          type: string
          format: uuid
        StoryId:
          type: string
          format: uuid
        ChapterNumber:
          type: integer
          format: int32
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
paths:
...
  /api/user/reading-history:
    get:
      summary: Lấy lịch sử đọc của người dùng
      description: Trả về danh sách lịch sử đọc của người dùng hiện tại dựa trên JWT token. Yêu cầu đăng nhập.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Danh sách lịch sử đọc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReadingHistoryResponse'
        '401':
          description: Unauthorized — token missing hoặc invalid
        '500':
          description: Lỗi server khi lấy lịch sử đọc
...